LayoutTests/platform/ios/imported/w3c/web-platform-tests/custom-elements/reactions/HTMLInputElement-expected.txt

CONSOLE MESSAGE: line 14: TypeError: The HTMLInputElement.capture getter can only be used on instances of HTMLInputElement

Harness Error (FAIL), message = TypeError: The HTMLInputElement.capture getter can only be used on instances of HTMLInputElement
